wip: 拓展函数生成逻辑

+++ example/src/main/kotlin/com/github/llh4github/jimmerhelper/example/dto/base.kt
@@ -1,5 +1,5 @@
 package com.github.llh4github.jimmerhelper.example.dto
 
 interface BaseDto {
-    val id: Int
+    val id: Long
 }
